requirements.txt
tensorflow==2.14.0
numpy==1.26.0
Pillow==10.0.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pydantic==2.5.2
python-multipart==0.0.6
